rnhp/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
Pillow==5.3.0
Markdown==3.0.1
